Rename and move ObservationDetailView and ObservationsListView
